pyproject.toml
[project]
name = "teste-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.10",
    "chromadb>=1.0.16",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "openai>=1.99.9",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
